pyproject.toml
[tool.poetry]
name = "kursovaya-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.9"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
